mcp-stack/host/requirements.txt
# Core dependencies
fastapi==0.95.2
httpx==0.24.1
pydantic>=1.10.0,<2.0.0
python-dotenv==1.0.0
uvicorn[standard]==0.22.0

# Data processing
numpy==1.24.3

# Logging
loguru==0.7.0

# GraphQL
gql[requests]==3.4.1

# Testing
pytest==7.4.0
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-httpx>=0.24.0

# Required by dependencies
anyio==3.7.1
click==8.1.3
h11==0.14.0
python-dateutil==2.8.2
sniffio==1.3.0
starlette==0.27.0
typing-extensions>=4.10.0
websockets==11.0.3
